requirements.txt
einops
matplotlib
numpy
pytorch-ignite
pynvml
PyYAML
tensorboard
torch
torchvision
tqdm
